Don't install sage/tests/cmdline.py with meson
It contains tests that are not relevant for the meson build

We extract the test_executable function to __init__.py, since it is still used in other files, and leave only the tests in cmdline.py
