+++ src/sage/repl/interpreter.py
@@ -722,7 +722,7 @@ def get_test_shell():
 
     Check that :issue:`14070` has been resolved::
 
-        sage: from sage.tests.cmdline import test_executable
+        sage: from sage.tests import test_executable
         sage: cmd = 'from sage.repl.interpreter import get_test_shell; shell = get_test_shell()'
         sage: (out, err, ret) = test_executable(["sage", "-c", cmd])
         sage: out + err
