+++ src/sage/tests/meson.build
@@ -7,7 +7,6 @@ py.install_sources(
   'book_schilling_zabrocki_kschur_primer.py',
   'book_stein_ent.py',
   'book_stein_modform.py',
-  'cmdline.py',
   'combinatorial_hopf_algebras.py',
   'finite_poset.py',
   'functools_partial_src.py',
